change the data type in the model

+++ chsdi/models/vector/uvek.py
@@ -693,7 +693,7 @@ class FernWaermeAngebot(Base, Vector):
     __label__ = 'name'
     id = Column('bgdi_id', Integer, primary_key=True)
     name = Column('name', Unicode)
-    heatpotential = Column('heatpotential', Integer)
+    heatpotential = Column('heatpotential', Float)
     heat_supplier_category = Column('heat_supplier_category', Integer)
     bezeichnung_de = Column('bezeichnung_de', Unicode)
     bezeichnung_fr = Column('bezeichnung_fr', Unicode)
